requirements.txt
discord.py
python-dotenv
pytz
line-bot-sdk>=3.0.0